fix : 타입 수정

+++ backend/src/main/java/signiel/heartsigniel/model/member/Member.java
@@ -56,13 +56,13 @@ public class Member {
     private String job;
 
     @Column(name = "si_do")
-    private int siDo;
+    private Long siDo;
 
     @Column(name = "gu_gun")
-    private int guGun;
+    private Long guGun;
 
     @Column(name = "is_first")
-    private int isFirst;
+    private Long isFirst;
